hic -- changes to format, added "whole genome" calculation
git-svn-id: https://igv.googlecode.com/svn/trunk@1009 17392f64-ead8-4cea-ae29-09b3ab513800

+++ src/org/broad/igv/hic/data/ContactRecord.java
@@ -9,9 +9,9 @@ public class ContactRecord {
     private int blockNumber;
     private int x;
     private int y;
-    private short counts;
+    private int counts;
 
-    public ContactRecord(int block, int x, int bin2, short counts) {
+    public ContactRecord(int block, int x, int bin2, int counts) {
         this.blockNumber = block;
         this.x = x;
         this.y = bin2;
@@ -34,7 +34,7 @@ public class ContactRecord {
         return y;
     }
 
-    public short getCounts() {
+    public int getCounts() {
         return counts;
     }
 }
